pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "patchpal"
dynamic = ["version"]
description = "An agentic coding and automation assistant, supporting both local and cloud LLMs"
readme = "README.md"
authors = [
    { name = "PatchPal Contributors" }
]
license = "Apache-2.0"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "litellm<=1.82.6",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "ddgs>=1.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "prompt_toolkit>=3.0.0",
    "boto3",
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
    "python-pptx>=0.6.0",
    "tree-sitter-language-pack>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff==0.14.13",
    "pre-commit>=3.0.0",
]
mcp = [
    "mcp>=0.9.0",
]
claw = [
    "croniter>=2.0.0",
    "python-telegram-bot>=20.0",
]
docs = [
    "mkdocs>=1.5.0,<2.0",  # Pin to 1.x (Material for MkDocs incompatible with 2.0)
    "mkdocs-material>=9.5.0",
    "mkdocs-autorefs>=1.0.0",
    "mkdocstrings>=0.24.0",
    "mkdocstrings-python>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/amaiya/patchpal"
Repository = "https://github.com/amaiya/patchpal"
Issues = "https://github.com/amaiya/patchpal/issues"

[project.scripts]
patchpal = "patchpal.cli.interactive:main"
patchpal-autopilot = "patchpal.cli.autopilot:main"
patchpal-daemon = "patchpal.claw.daemon:main"
patchpal-mcp = "patchpal.cli.mcp:main"
patchpal-sandbox = "patchpal.cli.sandbox:main"

[tool.setuptools.packages.find]
include = ["patchpal*"]

[tool.setuptools.dynamic]
version = {attr = "patchpal.__version__"}

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["patchpal"]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
